+++ src/Client/Query/Encoders/SimpleEncoder.php
@@ -8,7 +8,8 @@ use Somnambulist\Components\ApiClient\Client\Query\Expression\CompositeExpressio
 use Somnambulist\Components\ApiClient\Client\Query\Expression\Expression;
 use Somnambulist\Components\ApiClient\Client\Query\Expression\ExpressionBuilder;
 
-use function array_merge;
+use function array_merge_recursive;
+use function in_array;
 use function is_null;
 
 /**
@@ -32,6 +33,13 @@ class SimpleEncoder extends AbstractEncoder
         self::PER_PAGE => 'per_page',
     ];
 
+    public function useNameForFiltersField(?string $key): self
+    {
+        $this->mappings[self::FILTERS] = $key;
+
+        return $this;
+    }
+
     protected function createFilters(?CompositeExpression $expression): array
     {
         if (is_null($expression)) {
@@ -42,18 +50,22 @@ class SimpleEncoder extends AbstractEncoder
 
         foreach ($expression->getParts() as $part) {
             if ($part instanceof Expression) {
-                if (ExpressionBuilder::EQ !== $part->operator) {
+                if (!in_array($part->operator, [ExpressionBuilder::EQ, ExpressionBuilder::IN])) {
                     throw QueryEncoderException::encoderDoesNotSupportOperator(self::class, $part->field, $part->operator);
                 }
 
-                $filters[$part->field] = $part->value;
+                $filters[$part->field] = $part->getValueAsString();
             } elseif($part instanceof CompositeExpression) {
                 if ($part->isOr()) {
                     throw QueryEncoderException::encoderDoesNotSupportNestedConditions(self::class, 'OR');
                 }
 
-                $filters = array_merge($filters, $this->createFilters($part));
+                $filters = array_merge_recursive($filters, $this->createFilters($part));
             }
+        }
+
+        if (!is_null($this->mappings[self::FILTERS])) {
+            return [$this->mappings[self::FILTERS] => $filters];
         }
 
         return $filters;